requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4

# LLM Providers
groq==0.15.0
google-genai==1.5.0

# Memory
chromadb==0.6.3
sentence-transformers==3.4.1

# Voice (Phase 2)
edge-tts==7.2.7
numpy==2.2.1

# Utilities
aiofiles==24.1.0
jinja2==3.1.5
python-multipart==0.0.20
websockets==14.2
